requirements.txt
google-generativeai==0.8.4
pillow==11.1.0
streamlit==1.42.2
streamlit-option-menu==0.4.0